feat(self-serve-ds): move Directory Sync mutations onto the DirectorySync resource
