+++ packages/clerk-js/src/core/resources/DirectorySync.ts
@@ -1,19 +1,27 @@
 import type {
+  ClerkPaginatedResponse,
+  DeletedObjectJSON,
+  DeletedObjectResource,
   DirectorySyncJSON,
   DirectorySyncJSONSnapshot,
   DirectorySyncProvider,
   DirectorySyncResource,
   DirectorySyncUserJSON,
   DirectorySyncUserResource,
+  GetDirectorySyncUsersParams,
+  UpdateDirectorySyncParams,
 } from '@clerk/shared/types';
 
+import { convertPageToOffsetSearchParams } from '../../utils/convertPageToOffsetSearchParams';
 import { unixEpochToDate } from '../../utils/date';
 import { BaseResource } from './Base';
+import { DeletedObject } from './DeletedObject';
 
 export class DirectorySync extends BaseResource implements DirectorySyncResource {
   id!: string;
   name!: string;
-  enterpriseConnectionId: string | null = null;
+  organizationId!: string;
+  enterpriseConnectionId!: string;
   endpointUrl!: string;
   provider!: DirectorySyncProvider;
   enabled!: boolean;
@@ -23,10 +31,74 @@ export class DirectorySync extends BaseResource implements DirectorySyncResource
   createdAt: Date | null = null;
   updatedAt: Date | null = null;
 
-  constructor(data: DirectorySyncJSON | DirectorySyncJSONSnapshot | null) {
+  constructor(data: DirectorySyncJSON | DirectorySyncJSONSnapshot | null, organizationId: string) {
     super();
+    this.organizationId = organizationId;
     this.fromJSON(data);
   }
+
+  private get directoryPath(): string {
+    return `/organizations/${this.organizationId}/enterprise_connections/${this.enterpriseConnectionId}/directory`;
+  }
+
+  update = async (params: UpdateDirectorySyncParams): Promise<DirectorySyncResource> => {
+    const body: Record<string, string | boolean> = {};
+    if (params.enabled !== undefined) {
+      body.enabled = params.enabled;
+    }
+    if (params.attributeMapping !== undefined) {
+      body.attribute_mapping = JSON.stringify(params.attributeMapping);
+    }
+
+    const json = (
+      await BaseResource._fetch<DirectorySyncJSON>({
+        path: this.directoryPath,
+        method: 'PATCH',
+        body: body as any,
+      })
+    )?.response as unknown as DirectorySyncJSON;
+
+    return new DirectorySync(json, this.organizationId);
+  };
+
+  rotateToken = async (): Promise<DirectorySyncResource> => {
+    const json = (
+      await BaseResource._fetch<DirectorySyncJSON>({
+        path: `${this.directoryPath}/rotate_api_key`,
+        method: 'POST',
+      })
+    )?.response as unknown as DirectorySyncJSON;
+
+    return new DirectorySync(json, this.organizationId);
+  };
+
+  delete = async (): Promise<DeletedObjectResource> => {
+    const json = (
+      await BaseResource._fetch<DeletedObjectJSON>({
+        path: this.directoryPath,
+        method: 'DELETE',
+      })
+    )?.response as unknown as DeletedObjectJSON;
+
+    return new DeletedObject(json);
+  };
+
+  getUsers = async (
+    params?: GetDirectorySyncUsersParams,
+  ): Promise<ClerkPaginatedResponse<DirectorySyncUserResource>> => {
+    const res = await BaseResource._fetch({
+      path: `${this.directoryPath}/users`,
+      method: 'GET',
+      search: convertPageToOffsetSearchParams(params),
+    });
+
+    const payload = res?.response as unknown as ClerkPaginatedResponse<DirectorySyncUserJSON> | undefined;
+
+    return {
+      total_count: payload?.total_count ?? 0,
+      data: (payload?.data ?? []).map(row => new DirectorySyncUser(row)),
+    };
+  };
 
   protected fromJSON(data: DirectorySyncJSON | DirectorySyncJSONSnapshot | null): this {
     if (!data) {
@@ -35,7 +107,7 @@ export class DirectorySync extends BaseResource implements DirectorySyncResource
 
     this.id = data.id;
     this.name = data.name;
-    this.enterpriseConnectionId = data.enterprise_connection_id ?? null;
+    this.enterpriseConnectionId = data.enterprise_connection_id;
     this.endpointUrl = data.endpoint_url;
     this.provider = data.provider;
     this.enabled = data.enabled;
